requirements.txt
blinker==1.9.0
certifi==2025.1.31
click==8.1.8
colorama==0.4.6
Flask==3.1.0
itsdangerous==2.2.0
Jinja2==3.1.6
MarkupSafe==3.0.2
numpy==2.2.5
pynmea2==1.19.0
pyproj==3.7.1
pyserial==3.5
scipy==1.15.2
Werkzeug==3.1.3
